locations: Get Trash icon from the folder attributes
Some folders or mounted locations may use some icons that are different
from the standard ones, so instead of just using a standard fallback
icon we can try to fetch it from the file info parameters.

The trash is a simple case in which we can just use this information to
show the proper icon without having us to hardcode anything.

Helps with #1508